fix: Update directory structure to be coherent with the Firebase package name
The new package name is now com.github.se.project (instead of com.android.sample)
It also update some dependencies for the example tests that were missing due to the fact that we reuse the build.gradle.kts of the Bootcamp.
