+++ app/src/main/java/com/github/se/project/SimpleData.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.github.se.project
 
 import kotlin.math.sqrt
