+++ app/src/test/java/com/github/se/project/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.github.se.project
 
 import org.junit.Assert.*
 import org.junit.Test
